requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
textblob==0.18.0.post0
python-dotenv==1.0.1
numpy<1.23.0,>=1.16.5